third_party/blink/web_tests/fast/alignment/parse-align-items-expected.txt

This is a testharness.js-based test.
FAIL Test getting align-items set through CSS. assert_equals: alignItems is not what is should. expected "last baseline" but got "normal"
PASS Test initial value of align-items through JS
PASS Test getting and setting align-items through JS
PASS Test bad combinations of align-items
PASS Test the value 'initial'
PASS Test the value 'initial' for grid containers
PASS Test the value 'initial' for flex containers
PASS Test the value 'inherit'
Harness: the test ran to completion.
